+++ common-util/api/leaderboard.js
@@ -10,12 +10,12 @@ export const getLeaderboardList = async () => {
   const response = await TileDocument.load(ceramic, process.env.NEXT_PUBLIC_STREAM_ID);
   const users = get(response, 'content.users') || [];
 
-  const usersList = users.filter((e) => !!e.wallet_address)
-    .filter((e) => e.points !== 0)
+  const usersList = users
+    .filter((user) => user.points !== 0 && (!!user.twitter_id || !!user.wallet_address))
     .map((user, index) => ({
       ...user,
       // adding a unique key to each user
-      rowKeyUi: `${user.wallet_address}-${user.discord_id}-${index}`,
+      rowKeyUi: [user.wallet_address, user.twitter_id, user.discord_id, index].filter(Boolean).join('-'),
     }));
 
   return usersList;
